Resolve recipe linking issues in docs
Fixes https://github.com/openrewrite/rewrite-spring/issues/601
