+++ src/main/resources/META-INF/rewrite/properties-to-kebab-case.yml
@@ -1,0 +1,30 @@
+#
+# Copyright 2024 the original author or authors.
+# <p>
+# Licensed under the Moderne Source Available License (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# <p>
+# https://docs.moderne.io/licensing/moderne-source-available-license
+# <p>
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+---
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.spring.PropertiesToKebabCase
+displayName: Normalize Spring properties to kebab-case
+description: >-
+  Normalize Spring properties to use lowercase and hyphen-separated syntax. 
+  For example, changing `spring.main.showBanner` to `spring.main.show-banner`. 
+  With [Spring's relaxed binding](https://docs.spring.io/spring-boot/docs/2.5.6/reference/html/features.html#features.external-config.typesafe-configuration-properties.relaxed-binding), 
+  `kebab-case` may be used in properties files and still be converted to configuration beans. 
+  Note, an exception to this is the case of `@Value`, which is match-sensitive. For example, `@Value("${anExampleValue}")` will not match `an-example-value`. 
+  [The Spring reference documentation recommends using `kebab-case` for properties where possible](https://docs.spring.io/spring-boot/docs/2.5.6/reference/html/features.html#features.external-config.typesafe-configuration-properties.relaxed-binding).
+recipeList:
+  - org.openrewrite.java.spring.PropertiesToKebabCaseYaml
+  - org.openrewrite.java.spring.PropertiesToKebabCaseProperties
