+++ src/testWithSpringBoot_2_4/java/org/openrewrite/java/spring/boot2/MigrateDatabaseCredentialsTest.java
@@ -17,6 +17,7 @@ package org.openrewrite.java.spring.boot2;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
+import org.openrewrite.config.Environment;
 import org.openrewrite.test.RecipeSpec;
 import org.openrewrite.test.RewriteTest;
 
@@ -27,7 +28,7 @@ class MigrateDatabaseCredentialsTest implements RewriteTest {
 
     @Override
     public void defaults(RecipeSpec spec) {
-        spec.recipe(new MigrateDatabaseCredentials());
+        spec.recipeFromResources("org.openrewrite.java.spring.boot2.MigrateDatabaseCredentials");
     }
 
     @DocumentExample
